Fix virtual route layout
