+++ templates/skeleton/app/layout.tsx
@@ -1,0 +1,46 @@
+import {useNonce, Analytics} from '@shopify/hydrogen';
+import {
+  Links,
+  Meta,
+  Scripts,
+  useRouteLoaderData,
+  ScrollRestoration,
+  Outlet,
+} from '@remix-run/react';
+import resetStyles from '~/styles/reset.css?url';
+import appStyles from '~/styles/app.css?url';
+import {PageLayout} from '~/components/PageLayout';
+import { RootLoader } from './root';
+
+export default function Layout() {
+  const nonce = useNonce();
+  const data = useRouteLoaderData<RootLoader>('root');
+
+  return (
+    <html lang="en">
+      <head>
+        <meta charSet="utf-8" />
+        <meta name="viewport" content="width=device-width,initial-scale=1" />
+        <link rel="stylesheet" href={resetStyles}></link>
+        <link rel="stylesheet" href={appStyles}></link>
+        <Meta />
+        <Links />
+      </head>
+      <body>
+        {data ? (
+          <Analytics.Provider
+            cart={data.cart}
+            shop={data.shop}
+            consent={data.consent}
+          >
+            <PageLayout {...data}><Outlet /></PageLayout>
+          </Analytics.Provider>
+        ) : (
+          <Outlet />
+        )}
+        <ScrollRestoration nonce={nonce} />
+        <Scripts nonce={nonce} />
+      </body>
+    </html>
+  );
+}
